--- cypress.json
@@ -1,10 +1,0 @@
-{
-  "experimentalSessionAndOrigin": true,
-  "trashAssetsBeforeRuns": true,
-  "testFiles": [
-    "tests/setup/**/*.spec.ts",
-    "tests/pages/**/*.spec.ts",
-    "tests/navigation/**/*.spec.ts"
-  ],
-  "defaultCommandTimeout": 60000
-}
