--- cypress/integration/po/components/checkbox-input.po.ts
@@ -1,20 +1,0 @@
-import ComponentPo from '@/cypress/integration/po/components/component.po';
-
-export default class CheckboxInputPo extends ComponentPo {
-  /**
-   * Click on the checkbox input button
-   * @returns
-   */
-  set(): Cypress.Chainable {
-    return this.input().click({ force: true });
-  }
-
-  /**
-   * Return the checkbox input button from a given container
-   * @returns
-   */
-  private input(): Cypress.Chainable {
-    return this.self()
-      .find('.checkbox-container');
-  }
-}
