--- cypress/integration/po/side-bars/product-side-nav.po.ts
@@ -1,30 +1,0 @@
-import ComponentPo from '@/cypress/integration/po/components/component.po';
-
-export default class ProductNavPo extends ComponentPo {
-  constructor() {
-    super('.side-nav');
-  }
-
-  /**
-   * Get all navigation accordion groups
-   * @returns {Cypress.Chainable}
-   */
-  groups(): Cypress.Chainable {
-    return this.self().find('.accordion.has-children');
-  }
-
-  /**
-   * Get all the expanded accordion groups
-   * @returns
-   */
-  expandedGroup(): Cypress.Chainable {
-    return this.self().find('.accordion.expanded');
-  }
-
-  /**
-   * Get all the visible child links
-   */
-  visibleNavTypes(): Cypress.Chainable {
-    return this.self().find('.accordion.expanded li.nav-type>a');
-  }
-}
