--- cypress/integration/tests/setup/rancher-setup.spec.ts
@@ -1,46 +1,0 @@
-import { RancherSetupPagePo } from '@/cypress/integration/po/pages/rancher-setup.po';
-import { RancherSetupAuthVerifyPage } from '@/cypress/integration/po/pages/rancher-setup-auth-verify.po';
-
-describe('Rancher setup', () => {
-  it('Requires initial setup', () => {
-    cy.visit('');
-
-    new RancherSetupPagePo().hasInfoMessage();
-    cy.url().should('equal', `${ Cypress.config().baseUrl }/auth/login`);
-  });
-
-  it('Set initial Docker bootstrap password and admin password as single session', () => {
-    RancherSetupPagePo.goTo(); // Needs to happen before the page element is created/located
-    cy.intercept('POST', '/v3-public/localProviders/local?action=login').as('bootstrapReq');
-    const rancherSetup = new RancherSetupPagePo();
-
-    rancherSetup.canSubmit()
-      .should('eq', true);
-    rancherSetup.password().set(Cypress.env('bootstrapPassword'));
-    rancherSetup.canSubmit()
-      .should('eq', true);
-    rancherSetup.submit();
-
-    cy.wait('@bootstrapReq').then((login) => {
-      expect(login.response?.statusCode).to.equal(200);
-      cy.url().should('equal', `${ Cypress.config().baseUrl }/auth/setup`);
-    });
-
-    cy.intercept('PUT', '/v1/userpreferences/*').as('firstLoginReq');
-
-    const rancherSetupAuthVerify = new RancherSetupAuthVerifyPage();
-
-    rancherSetupAuthVerify.canSubmit()
-      .should('eq', false);
-    rancherSetupAuthVerify.termsAgreement().set();
-    rancherSetupAuthVerify.canSubmit()
-      .should('eq', true);
-    rancherSetupAuthVerify.submit();
-
-    // TODO: This assertion is commented as it started to fail after rebasing and cannot be corrected as it's not possible to run Rancher locally
-    // cy.wait('@firstLoginReq').then((login) => {
-    //   expect(login.response?.statusCode).to.equal(200);
-    //   cy.url().should('equal', `${ Cypress.config().baseUrl }/home`);
-    // });
-  });
-});
