+++ package.json
@@ -125,8 +125,8 @@
     "@babel/plugin-proposal-optional-chaining": "^7.14.5",
     "@babel/plugin-proposal-private-property-in-object": "^7.14.5",
     "@babel/preset-typescript": "^7.15.0",
-    "@cypress/vue": "^3.1.1",
-    "@cypress/webpack-dev-server": "^1.8.4",
+    "@cypress/vue": "^4.0.0",
+    "@cypress/webpack-dev-server": "^2.0.0",
     "@nuxt/types": "2.14.6",
     "@nuxt/typescript-build": "^2.1.0",
     "@nuxtjs/eslint-config-typescript": "^6.0.1",
@@ -153,7 +153,7 @@
     "core-js": "^3.20.3",
     "css-loader": "4.3.0",
     "csv-loader": "^3.0.3",
-    "cypress": "^9.6.0",
+    "cypress": "^10.3.1",
     "eslint": "^7.32.0",
     "eslint-config-standard": ">=12.0.0",
     "eslint-import-resolver-node": "0.3.4",
